requirements.txt
torch
transformers
openai-whisper
webrtcvad
sounddevice
numpy
textblob
TTS
simpleaudio
gradio
scipy
pyyaml
requests